requirements.txt
cloudpickle==3.1.1
cv_bridge==1.13.0.post0
dynamixel_sdk==3.7.31
hydra-core==1.3.2
ipdb==0.13.13
matplotlib==3.5.1
numpy==1.21.5
opencv_python==4.10.0.84
Pillow==9.0.1
pybullet==3.2.7
pynput==1.8.1
PyOpenGL==3.1.9
pytest==6.2.5
pyudev==0.24.3
PyYAML==6.0.2
pyzmq==26.4.0
ray==2.44.1
scipy==1.8.0
setuptools==59.6.0
termcolor==3.1.0
torch==2.7.1
tqdm==4.67.1
xarm==0.0.4
